+++ pigeons/flutter_zebra_capture_pigeon.dart
@@ -66,6 +66,7 @@ enum CaptureCapabilityStatus {
   unavailable,
   disconnected,
   connecting,
+  verifying,
   connected,
   error,
 }